Enable scene synchronization without an client Controller.
This commit allows to client to receive the server snapshot sync even
when the local controller doesn't exist.

The InputID generated by the controller is now optinal, when the
Controller doesn't exist; in such case the server snapshot is
immediately applied while the rewinding mechanism is entirely disabled.

It was necessary to modify the snapshot format: now the `InputID` is
stored separately from the controller `NodeData` and moved at the top of the array.

Receiving a corrupted snapshot data was fatal for the client, as the
corrupted data were leaking into the buffer, leaving the client unable
to function properly. This commit fix that by making sure a separate
buffer is used during the snapshot parsing.

The snapshot parsing algorithm was initializing the variables ONLY if a
variable was specified on the received snapshot. This was causing a
crash if the snapshot contained an empty NodeData buffer. This got fixed
by making sure the variables are initialized on the snapshot NodeData
parsing.